Yuji Chai
I have added the navigation part to the website. So that when we are going through it, there should not be so difficult as now. I also change the style for all the titles. So that we do not just the same as all the other part. I also added some other titles for clearance. The original code are written in Markdown, which is totally unfamiliar to me. I have tested all my codes they should just be fine. There might be some problems with the old configuration. If you want me to do some thing more. Just tell me right way. Thank you!